requirements.txt
chromadb==1.4.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
Requests==2.32.5
smolagents==1.24.0
smolagents>=1.24.0
chromadb>=1.4.1
sentence-transformers>=5.2.0
python-dotenv>=1.2.1
torch>=2.9.0
transformers>=4.57.0
huggingface-hub>=0.36.0